backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
djangorestframework
psycopg[binary]
drf-spectacular
djangorestframework
djangorestframework-simplejwt
django-cors-headers
